+++ CHANGES.md
@@ -2,8 +2,9 @@
 
 ## 1.43.1 (2024-11-22)
 
+- ENH: Add a environment variable to update SNAP log level ([#25](https://github.com/sertit/sertit-utils/issues/25))
+- FIX: Better management of default SNAP values (called in `snap.get_gpt_cli`, not during initialization)
 - FIX: Create threaded cluster by default in `rasters.get_or_create_dask_client`
-
 
 ## 1.43.1 (2024-11-22)